# Optimize `normalizeSearchTerm` function